pyproject.toml
[project]
name = "dlt-duck"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
# Pinned for DLT
requires-python = "==3.10.13"
dependencies = [
    "dlt[duckdb] ==0.4.12",
    "jupyter>=1.1.1",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.1",
]

[tool.uv]
dev-dependencies = ["pytest >=8.1.1,<9"]
